FEATURE(event-agent): use the new replication filter
 ##### SUMMARY

Use the new replication filter ("replicate") which is able to check in
the account service if the replication is enabled for the container emitting
each event. In case the replication is not enabled, the event is dropped.

By default, this new behavior is disabled, and each event will be forwarded
to the replication queue. Set check_replication_enabled to true to
enable the new behavior.

These new parameters have also been implemented:
- connection_timeout: connection timeout to the account service.
- read_timeout: read timeout for the account service.
- cache_duration: how long to keep the replication state before checking
  again in the account service.
- cache_size: how many containers to keep in the cache.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION
The new filter will be available in oio-sds starting from version 5.4.0.
The old generic "notify" filter will remain available.